@@ -1,12 +1,15 @@
 {
 	"name": "derekjohnston.ca",
-	"version": "1.0.0",
+	"version": "2.0.0",
 	"private": true,
+	"type": "module",
 	"scripts": {
-		"dev": "nuxt",
 		"build": "nuxt build",
+		"dev": "nuxt dev",
 		"start": "nuxt start",
 		"generate": "nuxt generate",
+		"preview": "nuxt preview",
+		"postinstall": "nuxt prepare",
 		"lint:js": "eslint --ext \".js,.ts,.vue\" --ignore-path .gitignore .",
 		"lint:prettier": "prettier --check .",
 		"lint": "yarn lint:js && yarn lint:prettier",
@@ -18,37 +21,41 @@
 		"hooks": "yarn lintfix"
 	},
 	"dependencies": {
-		"@fortawesome/fontawesome-svg-core": "^6.2.1",
-		"@fortawesome/free-brands-svg-icons": "^6.2.1",
+		"@fortawesome/fontawesome-svg-core": "^6.5.1",
+		"@fortawesome/free-brands-svg-icons": "^6.5.1",
+		"@fortawesome/free-regular-svg-icons": "^6.5.1",
 		"@fortawesome/free-solid-svg-icons": "^6.5.1",
-		"@fortawesome/vue-fontawesome": "^2.0.10",
+		"@fortawesome/vue-fontawesome": "^3.0.6",
 		"@nuxtjs/axios": "^5.13.6",
 		"@nuxtjs/fontawesome": "^1.1.2",
 		"@vue/language-plugin-pug": "^1.8.27",
 		"axios": "^1.6.7",
-		"bootstrap": "^4.6.2",
-		"bootstrap-vue": "^2.22.0",
-		"core-js": "^3.25.3",
-		"nuxt": "^2.15.8",
+		"bootstrap": "^5.3.3",
+		"bootstrap-vue": "^2.23.1",
+		"nuxt": "^3.10.3",
 		"pug": "^3.0.2",
 		"pug-plain-loader": "^1.1.0",
-		"vue": "^2.7.10",
-		"vue-server-renderer": "^2.7.10",
-		"vue-template-compiler": "^2.7.10"
+		"vite-plugin-pug": "^0.3.2",
+		"vue": "^3.4.19",
+		"vue-router": "^4.3.0"
 	},
 	"devDependencies": {
-		"@babel/eslint-parser": "^7.19.1",
-		"@nuxt/types": "^2.15.8",
-		"@nuxt/typescript-build": "^2.1.0",
-		"@nuxtjs/eslint-config-typescript": "^11.0.0",
-		"@nuxtjs/eslint-module": "^3.1.0",
-		"@nuxtjs/google-analytics": "^2.4.0",
-		"eslint": "^8.24.0",
-		"eslint-config-prettier": "^8.5.0",
+		"@nuxt/eslint-config": "^0.2.0",
+		"@nuxtjs/eslint-module": "^4.1.0",
+		"@types/bootstrap": "^5.2.10",
+		"@typescript-eslint/eslint-plugin": "^7.1.1",
+		"@typescript-eslint/parser": "^7.1.1",
+		"@vue/language-plugin-pug": "^2.0.6",
+		"eslint": "^8.57.0",
+		"eslint-config-prettier": "^9.1.0",
 		"eslint-plugin-nuxt": "^4.0.0",
-		"eslint-plugin-vue": "^9.5.1",
+		"eslint-plugin-prettier": "^5.1.3",
+		"eslint-plugin-vue": "^9.22.0",
 		"husky": "^9.0.11",
 		"pinst": "^3.0.0",
-		"prettier": "3.2.5"
+		"prettier": "^3.2.5",
+		"sass": "^1.71.1",
+		"sass-loader": "^14.1.1",
+		"typescript": "^5.4.2"
 	}
 }
